feat: ajustes cadastro e wip telas principais

+++ app/src/main/java/br/com/fiap/mercadoverde/utils/FormatCurrency.kt
@@ -3,7 +3,7 @@ package br.com.fiap.mercadoverde.utils
 import java.text.NumberFormat
 import java.util.Locale
 
-fun formatCurrency(value: Float): String {
+fun formatCurrency(value: Double): String {
     val format = NumberFormat.getCurrencyInstance(Locale("pt", "BR"))
     return format.format(value)
 }
